t3spi and struct union stuff
t3spi has been stripped to only the essential but is not going to be used. There is going to be a function at the bottom of main that initializes the spi. It is called in the setup function. I also wrote the typedef code for the register struct and the union types. Need to understand the details of this better.