We figured out that the double buffered SPI hardware cannot be worked around
There is a one byte lag in the spi message due to the offset of the push register from the shift register. The shift register as it turns out is loaded with the value in the SPI0_PUSHR_SLAVE. This value is loaded in the interrupt prior. So the outgoing message is phase shifted in a detrimental manner. The master will recieve the required read data after TWO sent bytes. It is my current suspicion that I the write messages are behaving similarly. Those are writing in a phase lagged manner as well.