Got register buffer working for the isr
Was able to make a copy of registers union called registers_buf that the isr will access instead of the actual registers. This way data is protected while the transfer is occuring. The registers are copied into buffer before every message and pulled out of buffer at the end of every message