Just Before Testing/Benchmarking Setup
This is currently in a state where it will print the message number, registers and register address, the time at the beginning and end of a register to register buffer transfer and the TX0 Buffer registers. The printed data may or may not be representative of what is actually happening in the registers. Not certain but will test now.