FINALLY DEBUGGED
Finally got this sucker working with the python script again.The print command sends a 1 to the correct address and it causes the firmware to print the registers outside of the isr. The registers print correctly corresponding to the previous write message that was sent. When not printing, the code does not totally shit the bricks and stop sending back the right stuff. Officially time to test with Simulink...NO PRINTING DURING THIS