IMU Data Type Stuff Success
Successfully implemented more direct way of getting IMU data from the BNO055. There was a private function in the adafruit library that grabs the I2C data directly. The other fucntions that are public would use this 16 bit message and convert it into a double at the proper 1 to 1 scaling. I bypassed this and then scaled by the proper factor in simulink to get the double precision. This way a 16 bit integer rather than an 8 bit double could be sent over spi to the raspberry pi