Moved init IMU code back, commented out that register, and added reset register
The IMU code would not work for some reason when not run in the context of the setup function. I moved it back and it worked again. Commented out the init_imu register stuff as a result and added a reset teensy register in the same spot so I dont have to redo all the registers on the Simulink side...It's a very big flaw that I am forced to ask for each register in a separate message. Need to figure out that memory allocation issue.