Added in reset imu again and added lots of comments for documentation
The IMU reset code was added in and is working. The reset code is the same as the initialization code. Began going through comments throroughly again. Rewording things and adding more comments to help with the firmware documentation. Beginning to ONLY comment, need to stop making changes. Oh yeah, and I made another change prior to this, I added in some initialization code for the timing variables so that the start variables are initialized only once at the beginning of loop(). Also moved the current time variables to the stack within loop(). Dont need them to be in RAM.