More comments
Added stuff to the spi message description. Need to move this section down to the interrupt service routines section and have that focus on the structure of the messages. Then I need to add a new description to the spi task. That one will focus on the sensor data gathering, the commands, and the actuations. Then I need to add a description to the setup functions and all RAM memory allocation/instantiation and initialization stuff up above setup() and loop().