Attempting to get Torque mode working
Copied over the startup and update velocity functions and changed what I believed was necessary to get the controllers configured for torque mode. New torque mode setup and update torque setpoint functions are included in this commit. They were tested and do not appear to be working. Wheels do not spin up when I squeeze the trigger in Manual mode Simulink model.