Added SDO filter to set torque mode function
Decided that all configuration functions should be selfish and ifgnore all other incoming CAN messages that are not responses to the messages they send out. This is not good practice in general, but for our purposes, this is ok. When the startup procedure is happening, each step is pre-defined and no other data should be important between steps in the procedure. So I made the torque configuration function filter out the SDO confirmation message form any junk that may be leftover in the CAN read buffer. I still need to incorporate wether the operating mode was changed or not in the functions returned variable.