More simulink changes and torque test file
Added in sign change for wheel velocities and changed the constant torque input. Made a torque test file that has two torque step inputs spaced out in time. Want to see the velocity response and get a feel for it. Is it approximately first order? What is its approximate time constant even if higher order.