Ran Torque tests
It has become apparent that I need to implement quick stop immediately. Some kind of state machine in Simulink that sends the quick stop command over spi if the user lets go of the throttle (dead switch) and waits for the system to reset or something. Need to write the firmware for this as well. Quick stop command and then back to the operation enabled state when the throttle is pressed again. NEED TO SEE HOW THIS AFFECTS THE DEAD SWITCH ON THE TEENSY