Major Simulink Chagnes
Changed logic so that entire enabled subsystem is turned on as a response to the dead switch being activated. This prevents integral wind-up and essentially resets the entire control system every time the trigger is pressed. Now sensor feedback for viewing or data storage purposes is outside of the subsystem and gathering constantly. This prevents feedback from going dormant when the control system and actuation subsystem is inactive.