dead_switch state machine change and CAN changes
Added a requirement to the dead switch state machine that the motor controller start up code has been run prior. This is done with the MC_on_flag Later, shutdown code will lower the MC_on_flag and the code wont run but this has not been implemented yet.

The CAN message filtering code was moved into a function at the bottom of the script. I started implementing SDO interpretting but only got so far as to figure out what the cases needed to be. There will need to be a function within the SDO switch case that further interprets the data of the switch case and updates the registers and flags accordingly.