Fixed packing issue with reg map and updated simulink. Also CAN stuff
Fixed the issue where the register union values would be strangely spaced out by the compiler. Added in a struct and union attribute (packed) to prevent this from happening so that the things I wanted to do with the CAN error and statuswords could be done. Added in request_error_registers function to the CAN driver (also renamed the ProgreSSIV_MC_driver to ProgreSSIV_CAN_driver - going back on old change). Implemented some dynamic SDO read requests for the errors but I'm not certain it's working properly yet.

Drastically changed the simulink SPI communication blocks to reflect the changes to the Teensy register union and struct.