MORE DEBUGGING CAN TEST CODE
The issue is persistent and UNPREDICTABLE. Seemingly harmless things like a delay whithin a while loop or adding in a for loop after the CAN code or something will cause the SDO recieve messages to time out (all or some or none). There is just no indication why this is occuring and I've been at it for 16+ hours. Going to try moving to a  further branch of FLEXCAN driver done by someone named Collin. Hopefully this fixes the issues, I'm not sure.