Got state machine integrated with main firmware and debugged to a mostly working state
Got the state machine integrated with the main Teensy firmware and debugged a lot of major issues. I made note of potential improvements to the startup code and some CAN buffer overload preventing features. Statuswords still acting funny in Simulink even after I added a logical and with the bitmask for the important bits.