Making good progress on the new architecture
Forgot to mention in last commit that I made a shared structs file that type defines the structs of data to be shared between tasks. No compile errors in MC_state_machine and still working on compile errors in SPI_task.

The registers_buf in spi0_isr has been replaced with the main registers union and I'm just going to make sure that SPI task doesnt update the registers or read from them if an SPI transfer is occuring.