Implementing SPI task - compiler issues and ISR dilemnas
The code is not currently compiling - issues with volatile types not being compatible with their non-volatile counterparts...may just have all data associated with SPI writes be volatile.

Having a dilemna with the SPI interrupt service routines where I cant decide how to prevent writing the low byte of a value with old SPI register data and the high byte of it with new SPI register data (for example). Have some ideas but have yet to implement them...and they all seem to have some flaws